+++ base/src/api/agent/mod.rs
@@ -9,7 +9,6 @@ use replicante_util_actixweb::TracingMiddleware;
 pub mod info;
 pub mod shards;
 
-use self::info::AgentInfo;
 use self::info::DatastoreInfo;
 use self::shards::Shards;
 
@@ -26,7 +25,6 @@ pub fn configure_app(
 ) {
     let tracer = Arc::clone(&context.tracer);
     APIRoot::UnstableAPI.and_then(flags, |root| {
-        let agent_for_agent = Arc::clone(&agent);
         let agent_for_datastore = Arc::clone(&agent);
         let agent_for_shards = agent;
         let cluster_display_name_override = context.config.cluster_display_name_override.clone();
@@ -36,7 +34,7 @@ pub fn configure_app(
                     context.logger.clone(),
                     Arc::clone(&tracer),
                 ))
-                .route(web::get().to(move || AgentInfo::new(Arc::clone(&agent_for_agent)))),
+                .route(web::get().to(info::agent)),
         );
         app.service(
             root.resource("/info/datastore")